requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-common-io==1.5.2
apache-airflow-providers-common-sql==1.25.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.1
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.2
apache-airflow-providers-sqlite==4.0.1
apispec==6.8.1
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
ChatterBot==1.2.6
chatterbot-corpus==1.2.2
click==8.1.8
clickclick==20.10.2
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==43.0.3
cymem==2.0.11
daphne==4.2.0
databases==0.9.0
Deprecated==1.2.18
dill==0.3.9
Django==4.2.20
djangorestframework==3.16.0
dnspython==2.7.0
docx==0.2.4
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-jwt-auth==0.5.0
feedparser==6.0.11
Flask==3.1.1
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2025.3.2
google-re2==1.1.20240702
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.10.0
limits==4.7.2
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.2
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mathparse==0.1.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
numpy==2.2.4
openpyxl==3.1.5
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-exporter-otlp-proto-http==1.32.0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
ordered-set==4.1.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pendulum==3.0.0
pika==1.3.2
pillow==11.1.0
pipx==1.7.1
platformdirs==4.3.7
pluggy==1.5.0
preshed==3.0.9
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==1.7.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-argparse==1.7.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9.1
service-identity==24.2.0
setproctitle==1.3.5
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejwt==2.0.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
thinc==8.3.6
time-machine==2.16.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uc-micro-py==1.0.3
ujson==5.10.0
universal_pathlib==0.2.6
urllib3==2.4.0
userpath==1.9.2
uv==0.6.16
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.19.0
zipp==3.21.0
zope.interface==7.2
